▷  Adding & ajusting functions
Created the select_penalty function that will find the ribosomal penalty score that minimises the PCA distance of the artificial cells to the nearest true cell neighbour (dTNN).

This is a precursor to the detect_damage function that uses simulation to predict what the input cells look like once they have lost cytoplasmic RNA. It will then match each true cell to the artificial damaged level to which it has the highest proportion of nearest neighbours (pANNs).