► Major computation adjustments
Previously, the detect_damage function depended on sets of artificial cell with damage profiles ranging from 0 to 1. This samples from all cells and has the chance of muddying the true damaged cell profiles.

Instead of this, a clustering component is introduced. This involves creating artificial cells for each cluster found using the standard Seurat clustering workflow with a low resoultion (r = 0.1). This assumes that each cluster is a distinct population and creates a damaged profile for the population with a high level of RNA loss (i.e. damaged cells).

The new method will cluster the cells based on gene expression to identify the different populations. Each population will be used as input for damage simulation, but only modelling extreme damage, between 0.65 and 1 percent of cytoplasmic RNA lost. The idea here is that we should get a good idea of what the pool of damaged cells for a mixed sample look like. When comparing these to the set of true cells, greater similarity to damaged cells than true cells indicates a cell has likely lost  the same RNA as the simulated damaged cells and is assumed to be itself damaged. Also, if a population is made up entirely of damaged cells created from the loss of cytoplasmic RNA in another population, this will be easily detected.